+++ src/main/java/com/micro/demo/controller/DirectorRestController.java
@@ -3,10 +3,10 @@ package com.micro.demo.controller;
 import com.micro.demo.configuration.Constants;
 import com.micro.demo.controller.dto.AprobarRechazarCambiosRequestDto;
 import com.micro.demo.controller.dto.AsignaturaDto;
-import com.micro.demo.controller.dto.AssignAsignaturasRequestDto;
-import com.micro.demo.controller.dto.AssignCompetenciaRequestDto;
-import com.micro.demo.controller.dto.AssignDocentesRequestDTO;
-import com.micro.demo.controller.dto.AssignTemasRequestDto;
+import com.micro.demo.controller.dto.request.assign.AssignAsignaturasRequestDto;
+import com.micro.demo.controller.dto.request.assign.AssignCompetenciaRequestDto;
+import com.micro.demo.controller.dto.request.assign.AssignDocentesRequestDTO;
+import com.micro.demo.controller.dto.request.assign.AssignTemasRequestDto;
 import com.micro.demo.controller.dto.CompetenciaDto;
 import com.micro.demo.controller.dto.RemoveAsignaturaRequestDto;
 import com.micro.demo.controller.dto.RemoveDocenteRequestDto;
