+++ src/main/java/com/micro/demo/controller/dto/request/assign/AssignTemasRequestDto.java
@@ -1,22 +1,19 @@
-package com.micro.demo.controller.dto;
+package com.micro.demo.controller.dto.request.assign;
 
 import jakarta.validation.constraints.NotNull;
-
-import java.util.List;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.util.List;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class AssignDocentesRequestDTO {
+public class AssignTemasRequestDto {
     @NotNull
-    private Long asignaturaId;
-    @NotNull
-    private List<String> correoDocentes;
-
+    private Long unidadId;
+    private List<Long> temaIds;
 }
